refactor(test): 合并advanced_cond_test.go到cond_test.go
1. 将advanced_cond_test.go中的所有测试合并到cond_test.go:
   - TestToInterfacesFunction: 测试toInterfaces函数
   - TestConditionCoverage: 测试$raw、$and、$or命令和布尔条件
   - TestSimpleTypeToStrCoverage: 测试不同数据类型转换
   - TestFieldNameValidation: 测试字段名验证
   - TestTablePrefixAndQuoting: 测试表前缀和引号处理

2. 删除advanced_cond_test.go文件

3. 修复mock_client_test.go编译错误:
   - 添加TestModelWithTimestamps类型定义

测试结果:
- 所有测试通过 ✓
- 当前覆盖率: 32.8%
